Add code for identity matrix creation
Creates an n x n matrix equivalent to eye(n) in matlab.

Note that this uses functions from element_arithmetic (fill_matrix)
so we need to update the dependencies both in the include and
makefile. We also need to prevent multiple declarations, hence the
inclusion of the include guard in the base matrix_ops.h