Add code for solving linear equations
Adds functions for solving equations of the form Ax=b.
This utilises a previous static function: gauss
elimination with partial pivoting, in order to yield
the PLU matrices in order to solve the linear equations.

This commit simply provides the code for back-substitution
in order to do this.

Tests provided, but further tests are required to test
dependent and inconsistent equations