Adapt makefile to create libraries for all types with make all
Depending on the use case, the user may have use for different
types, and it is cumbersome to mandate a recompilation and
reinstallation each time the types change.

Therefore, make all now creates all libraries, and will
all be installed at the same time.

In addition, tests are now not created with make all, but now
with make tests. This is because tests can only be made with
a single defined type, where make all now makes libraries
for all types.